subject controller

+++ src/main/java/com/example/technopolis/controllers/SubjectController.java
@@ -2,9 +2,7 @@ package com.example.technopolis.controllers;
 
 import com.example.technopolis.dto.SubjectDto;
 import com.example.technopolis.model.Subject;
-import com.example.technopolis.model.User;
 import com.example.technopolis.services.SubjectRepository;
-import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,7 +14,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/subjects/{id}")
@@ -27,23 +29,42 @@ public class SubjectController {
 
     @GetMapping
     public ModelAndView getSubjects(@PathVariable long id) {
-        return new ModelAndView("index", Collections.singletonMap("subjects", repository.findAll()));
+        List<Subject> list = new ArrayList<>();
+        for (Subject subject : repository.findAll()) {
+            if (subject.getUser().getId() == id) {
+                list.add(subject);
+            }
+        }
+        return new ModelAndView("index", Collections.singletonMap("subjects", list));
     }
 
     @PutMapping
-    public Subject insertSubject(@RequestBody SubjectDto subjectDto,@PathVariable long id) {
-        return repository.save(new Subject(subjectDto.getInputName()));
+    public Subject insertSubject(@RequestBody SubjectDto subjectDto, @PathVariable long id) {
+        return repository.save(new Subject(subjectDto.getInputName(), id));
     }
 
     @PostMapping
-    public Subject updateSubject(@RequestBody SubjectDto subjectDto,@PathVariable long id) {
-        Subject subject = repository.findOne(subjectDto.getInputId());
-        subject.setName(subjectDto.getInputName());
-        return repository.save(subject);
+    public Subject updateSubject(@RequestBody SubjectDto subjectDto, @PathVariable long id) {
+        for (Subject subject : repository.findAll()) {
+            if (subject.getUser().getId() == id) {
+                if (Objects.equals(subject.getId(), subjectDto.getInputId())) {
+                    subject.setName(subjectDto.getInputName());
+                    return repository.save(subject);
+                }
+            }
+        }
+        throw new EntityNotFoundException();
     }
 
     @DeleteMapping
-    public void deleteSubject(@RequestBody SubjectDto subjectDto) {
-        repository.delete(subjectDto.getInputId());
+    public void deleteSubject(@RequestBody SubjectDto subjectDto, @PathVariable long id) {
+        for (Subject subject : repository.findAll()) {
+            if (subject.getUser().getId() == id) {
+                if (Objects.equals(subject.getId(), subjectDto.getInputId())) {
+                    repository.delete(subjectDto.getInputId());
+                }
+            }
+        }
+        throw new EntityNotFoundException();
     }
 }
